Renamed type.h to common.h and moved into core/common.h. Also renamed type_utils to expr_utils
