+++ src/core/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_library(Core STATIC
     utils/input.c
     utils/log.c
-    utils/type_utils.c
+    utils/expr_utils.c
     
     context/config.c
     context/context.c
