+++ src/core/common.h
