+++ src/core/execution/executor.c
@@ -3,10 +3,9 @@
 
 #include "executor.h"
 #include "core/context.h"
-#include "core/primitives/types.h"
 #include "simplify.h"
 
-#include "core/utils/type_utils.h"
+#include "core/utils/expr_utils.h"
 #include "core/utils/log.h"
 #include "core/context/environment.h"
