+++ src/core/parsing/codegen/ast.c
@@ -4,10 +4,9 @@
 #include "ast.h"
 #include "core/context.h"
 #include "core/parsing/parser_types.h"
-#include "core/primitives/types.h"
 #include "core/utils/log.h"
 #include "core/parsing/parser_utils.h"
-#include "core/utils/type_utils.h"
+#include "core/utils/expr_utils.h"
 
 #define DEFAULT_NODE_STACK_SIZE 10
