+++ src/core/parsing/parser.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "core/primitives/types.h"
+#include "core/common.h"
 
 
 typedef enum PARSER_RESULT {
