+++ src/core/parsing/parser_utils.c
@@ -4,11 +4,11 @@
 #include <string.h>
 
 #include "parser_utils.h"
+#include "core/common.h"
 #include "core/context.h"
 #include "core/parsing/parser_types.h"
 #include "core/utils/log.h"
-#include "core/primitives/types.h"
-#include "core/utils/type_utils.h"
+#include "core/utils/expr_utils.h"
 #include "core/primitives/integers.h"
