+++ src/core/utils/expr_utils.h
@@ -2,7 +2,7 @@
 
 #include <stdio.h>
 
-#include "core/primitives/types.h"
+#include "core/common.h"
 
 
 Expression *dummyExpression(ExpressionType type);
