+++ src/main.c
@@ -3,7 +3,7 @@
 #include <string.h>
 
 #include "core/context.h"
-#include "core/primitives/types.h"
+#include "core/common.h"
 #include "core/utils/log.h"
 #include "core/utils/input.h"
