+++ src/modules/algebra.c
@@ -1,5 +1,5 @@
 #include "algebra.h"
-#include "core/utils/type_utils.h"
+#include "core/utils/expr_utils.h"
 
 long long _gcd(long long a, long long b) {
     long long temp;
