+++ src/modules/algebra.h
@@ -1,7 +1,7 @@
 #ifndef ALGEBRA_H
 #define ALGEBRA_H
 
-#include "core/primitives/types.h"
+#include "core/common.h"
 
 
 long long _gcd(long long a, long long b);
